+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Fixed
 - Problems with cli not displaying capfinder version info
+- Fixed issue with API docs generation
 
 ## [0.2.6] - 2024-08-01
 ### Fixed