+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.8] - 2024-08-02
 ### Fixed
 - Fixed loading the entire inference dataset in memory
 
@@ -92,7 +94,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Basic skeleton of the package and tested it
 
-[Unreleased]: https://github.com/adnaniazi/capfinder/compare/0.2.7...master
+[Unreleased]: https://github.com/adnaniazi/capfinder/compare/0.2.8...master
+[0.2.8]: https://github.com/adnaniazi/capfinder/compare/0.2.7...0.2.8
 [0.2.7]: https://github.com/adnaniazi/capfinder/compare/0.2.6...0.2.7
 [0.2.6]: https://github.com/adnaniazi/capfinder/compare/0.2.5...0.2.6
 [0.2.5]: https://github.com/adnaniazi/capfinder/compare/0.2.4...0.2.5
@@ -111,3 +114,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [0.1.2]: https://github.com/adnaniazi/capfinder/compare/0.1.1...0.1.2
 [0.1.1]: https://pypi.org/manage/project/capfinder/release/0.1.1/
 [0.1.0]: https://pypi.org/manage/project/capfinder/release/0.1.0/
+